create new package tenants for tenant onboarding logic
