+++ tenants/parse.go
@@ -1,0 +1,157 @@
+package tenants
+
+import (
+	"flag"
+	"fmt"
+	"net/url"
+	"strings"
+	"time"
+
+	"github.com/observatorium/api/logger"
+	"github.com/observatorium/api/tracing"
+)
+
+//nolint:funlen
+func parseFlags() (config, error) {
+	var (
+		rawTLSCipherSuites      string
+		rawMetricsReadEndpoint  string
+		rawMetricsWriteEndpoint string
+		rawLogsReadEndpoint     string
+		rawLogsTailEndpoint     string
+		rawLogsWriteEndpoint    string
+		rawTracingEndpointType  string
+	)
+
+	cfg := config{}
+
+	flag.StringVar(&cfg.rbacConfigPath, "rbac.config", "rbac.yaml",
+		"Path to the RBAC configuration file.")
+	flag.StringVar(&cfg.tenantsConfigPath, "tenants.config", "tenants.yaml",
+		"Path to the tenants file.")
+	flag.StringVar(&cfg.debug.name, "debug.name", "observatorium",
+		"A name to add as a prefix to log lines.")
+	flag.IntVar(&cfg.debug.mutexProfileFraction, "debug.mutex-profile-fraction", 10,
+		"The percentage of mutex contention events that are reported in the mutex profile.")
+	flag.IntVar(&cfg.debug.blockProfileRate, "debug.block-profile-rate", 10,
+		"The percentage of goroutine blocking events that are reported in the blocking profile.")
+	flag.StringVar(&cfg.logLevel, "log.level", "info",
+		"The log filtering level. Options: 'error', 'warn', 'info', 'debug'.")
+	flag.StringVar(&cfg.logFormat, "log.format", logger.LogFormatLogfmt,
+		"The log format to use. Options: 'logfmt', 'json'.")
+	flag.StringVar(&cfg.internalTracing.serviceName, "internal.tracing.service-name", "observatorium_api",
+		"The service name to report to the tracing backend.")
+	flag.StringVar(&cfg.internalTracing.endpoint, "internal.tracing.endpoint", "",
+		"The full URL of the trace agent or collector. If it's not set, tracing will be disabled.")
+	flag.StringVar(&rawTracingEndpointType, "internal.tracing.endpoint-type", string(tracing.EndpointTypeAgent),
+		fmt.Sprintf("The tracing endpoint type. Options: '%s', '%s'.", tracing.EndpointTypeAgent, tracing.EndpointTypeCollector))
+	flag.Float64Var(&cfg.internalTracing.samplingFraction, "internal.tracing.sampling-fraction", 0.1,
+		"The fraction of traces to sample. Thus, if you set this to .5, half of traces will be sampled.")
+	flag.StringVar(&cfg.server.listen, "web.listen", ":8080",
+		"The address on which the public server listens.")
+	flag.StringVar(&cfg.server.listenInternal, "web.internal.listen", ":8081",
+		"The address on which the internal server listens.")
+	flag.StringVar(&cfg.server.healthcheckURL, "web.healthchecks.url", "http://localhost:8080",
+		"The URL against which to run healthchecks.")
+	flag.StringVar(&rawLogsTailEndpoint, "logs.tail.endpoint", "",
+		"The endpoint against which to make tail read requests for logs.")
+	flag.StringVar(&rawLogsReadEndpoint, "logs.read.endpoint", "",
+		"The endpoint against which to make read requests for logs.")
+	flag.StringVar(&cfg.logs.tenantHeader, "logs.tenant-header", "X-Scope-OrgID",
+		"The name of the HTTP header containing the tenant ID to forward to the logs upstream.")
+	flag.StringVar(&rawLogsWriteEndpoint, "logs.write.endpoint", "",
+		"The endpoint against which to make write requests for logs.")
+	flag.StringVar(&rawMetricsReadEndpoint, "metrics.read.endpoint", "",
+		"The endpoint against which to send read requests for metrics. It used as a fallback to 'query.endpoint' and 'query-range.endpoint'.")
+	flag.StringVar(&rawMetricsWriteEndpoint, "metrics.write.endpoint", "",
+		"The endpoint against which to make write requests for metrics.")
+	flag.StringVar(&cfg.metrics.tenantHeader, "metrics.tenant-header", "THANOS-TENANT",
+		"The name of the HTTP header containing the tenant ID to forward to the metrics upstreams.")
+	flag.StringVar(&cfg.metrics.tenantLabel, "metrics.tenant-label", "tenant_id",
+		"The name of the PromQL label that should hold the tenant ID in metrics upstreams.")
+	flag.StringVar(&cfg.tls.serverCertFile, "tls.server.cert-file", "",
+		"File containing the default x509 Certificate for HTTPS. Leave blank to disable TLS.")
+	flag.StringVar(&cfg.tls.serverKeyFile, "tls.server.key-file", "",
+		"File containing the default x509 private key matching --tls.server.cert-file. Leave blank to disable TLS.")
+	flag.StringVar(&cfg.tls.healthchecksServerCAFile, "tls.healthchecks.server-ca-file", "",
+		"File containing the TLS CA against which to verify servers."+
+			" If no server CA is specified, the client will use the system certificates.")
+	flag.StringVar(&cfg.tls.healthchecksServerName, "tls.healthchecks.server-name", "",
+		"Server name is used to verify the hostname of the certificates returned by the server."+
+			" If no server name is specified, the server name will be inferred from the healthcheck URL.")
+	flag.StringVar(&cfg.tls.minVersion, "tls.min-version", "VersionTLS13",
+		"Minimum TLS version supported. Value must match version names from https://golang.org/pkg/crypto/tls/#pkg-constants.")
+	flag.StringVar(&rawTLSCipherSuites, "tls.cipher-suites", "",
+		"Comma-separated list of cipher suites for the server."+
+			" Values are from tls package constants (https://golang.org/pkg/crypto/tls/#pkg-constants)."+
+			" If omitted, the default Go cipher suites will be used."+
+			" Note that TLS 1.3 ciphersuites are not configurable.")
+	flag.DurationVar(&cfg.tls.reloadInterval, "tls.reload-interval", time.Minute,
+		"The interval at which to watch for TLS certificate changes.")
+	flag.StringVar(&cfg.middleware.rateLimiterAddress, "middleware.rate-limiter.grpc-address", "",
+		"The gRPC Server Address against which to run rate limit checks when the rate limits are specified for a given tenant."+
+			" If not specified, local, non-shared rate limiting will be used.")
+	flag.IntVar(&cfg.middleware.concurrentRequestLimit, "middleware.concurrent-request-limit", 10_000,
+		"The limit that controls the number of concurrently processed requests across all tenants.")
+	flag.IntVar(&cfg.middleware.backLogLimitConcurrentRequests, "middleware.backlog-limit-concurrent-requests", 0,
+		"The number of concurrent requests that can buffered.")
+	flag.DurationVar(&cfg.middleware.backLogDurationConcurrentRequests, "middleware.backlog-duration-concurrent-requests", 1*time.Millisecond,
+		"The time duration to buffer up concurrent requests.")
+
+	flag.Parse()
+
+	metricsReadEndpoint, err := url.ParseRequestURI(rawMetricsReadEndpoint)
+	if err != nil {
+		return cfg, fmt.Errorf("--metrics.read.endpoint %q is invalid: %w", rawMetricsReadEndpoint, err)
+	}
+
+	cfg.metrics.readEndpoint = metricsReadEndpoint
+
+	metricsWriteEndpoint, err := url.ParseRequestURI(rawMetricsWriteEndpoint)
+	if err != nil {
+		return cfg, fmt.Errorf("--metrics.write.endpoint %q is invalid: %w", rawMetricsWriteEndpoint, err)
+	}
+
+	cfg.metrics.writeEndpoint = metricsWriteEndpoint
+
+	if rawLogsReadEndpoint != "" {
+		cfg.logs.enabled = true
+
+		logsReadEndpoint, err := url.ParseRequestURI(rawLogsReadEndpoint)
+		if err != nil {
+			return cfg, fmt.Errorf("--logs.read.endpoint is invalid, raw %s: %w", rawLogsReadEndpoint, err)
+		}
+
+		cfg.logs.readEndpoint = logsReadEndpoint
+	}
+
+	if rawLogsTailEndpoint != "" {
+		cfg.logs.enabled = true
+
+		logsTailEndpoint, err := url.ParseRequestURI(rawLogsTailEndpoint)
+		if err != nil {
+			return cfg, fmt.Errorf("--logs.tail.endpoint is invalid, raw %s: %w", rawLogsTailEndpoint, err)
+		}
+
+		cfg.logs.tailEndpoint = logsTailEndpoint
+	}
+
+	if rawLogsWriteEndpoint != "" {
+		cfg.logs.enabled = true
+
+		logsWriteEndpoint, err := url.ParseRequestURI(rawLogsWriteEndpoint)
+		if err != nil {
+			return cfg, fmt.Errorf("--logs.write.endpoint is invalid, raw %s: %w", rawLogsWriteEndpoint, err)
+		}
+
+		cfg.logs.writeEndpoint = logsWriteEndpoint
+	}
+
+	if rawTLSCipherSuites != "" {
+		cfg.tls.cipherSuites = strings.Split(rawTLSCipherSuites, ",")
+	}
+
+	cfg.internalTracing.endpointType = tracing.EndpointType(rawTracingEndpointType)
+
+	return cfg, nil
+}
